fixed apsuse and ptuse ephemeries folding bugs

+++ modules.nf
@@ -301,13 +301,80 @@ process dspsr_fold_ephemeris {
     """
 }
 
+process dspsr_fold_ephemeris_apsuse {
+    label 'fold_ephemeris'
+    container "${params.fold_singularity_image}"
+
+    input:
+    tuple val(files), val(target_name), val(beam_name), val(utc_start), path(ephemeris_file)
+    val(threads)
+    val(telescope)
+    val(subint_length)
+    val(bins)
+
+    output:
+    path "*.ar"
+
+    script:
+    """
+    # Extracting basename and filename without extension from ephemeris_file
+    ephemeris_basename=\$(basename ${ephemeris_file})
+    ephemeris_name_no_ext=\${ephemeris_basename%.*}
+
+    # Initialize an array to keep track of successfully processed files
+    declare -a successful_files
+
+    # Split the 'files' variable into an array based on space separator
+    IFS=' ' read -r -a input_files_array <<< "${files}"
+
+    # Loop through all files in the input_files_array
+    for input_file in "\${input_files_array[@]}"; do
+        # Extract the filename without its path and extension
+        filename=\$(basename "\$input_file")
+        filename_no_ext="\${filename%.*}"
+        output_filename="\${filename_no_ext}_\${ephemeris_name_no_ext}"
+
+        # Process each file with dspsr. If dspsr fails for a file, the script continues without exiting or throwing an error.
+        if dspsr -A -E ${ephemeris_file} -L ${subint_length} -b ${bins} -t ${threads} -k ${telescope} -e ar -O "\${output_filename}" "\$input_file" >& /dev/null; then
+            # If dspsr succeeds, add the path of the output .ar file to the list of successful files
+            successful_files+=("\${output_filename}.ar")
+
+        fi
+    done
+
+    # Check if there are any successful files to process further
+    final_output_name="${target_name}_${utc_start}_\${ephemeris_name_no_ext}.ar"
+    if [[ \${#successful_files[@]} -eq 1 ]]; then
+        # If there is exactly one successful file, rename it to match the final output name
+        mv "\${successful_files[0]}" "\$final_output_name"
+
+    elif [[ \${#successful_files[@]} -gt 1 ]]; then
+        # If there are multiple successful files, use psradd to combine them into one archive with the final output name
+        psradd -o "\$final_output_name" *.ar
+        # Cleanup the individual .ar files
+        
+        for file in "\${successful_files[@]}"; do
+            #This check is for the edge case when user requested subint length is smaller than tobs, dspsr runs successfully, but no outputfile is created!
+            if [ -e "\$file" ]; then
+                rm -rfv "\$file"
+            fi
+        done
+    else
+        echo "No files were successfully processed by dspsr."
+    fi
+    """
+}
+
+
+
+
 
 process dspsr_fold_ephemeris_ptuse {
     label 'fold_ephemeris'
     container "${params.fold_singularity_image}"
 
     input:
-    tuple val(file_path), val(target_name), val(beam_name), val(utc_start), path(ephemeris_file)
+    tuple val(file_path1), val(file_path2), val(file_path3), val(target_name), val(beam_name), val(utc_start), path(ephemeris_file)
     val(threads)
     val(telescope)
     val(subint_length)
@@ -319,11 +386,19 @@ process dspsr_fold_ephemeris_ptuse {
     script:
     """
     # Find at least one file that matches the wildcard pattern
-    found_file=\$(ls -v ${file_path} 2> /dev/null | head -n 1)
+    found_file1=\$(ls -v ${file_path1} 2> /dev/null | head -n 1)
+    found_file2=\$(ls -v ${file_path2} 2> /dev/null | head -n 1)
+    found_file3=\$(ls -v ${file_path3} 2> /dev/null | head -n 1)
 
-    if [[ -z "\$found_file" ]]; then
-        echo "Error: No matching files found."
-        exit 1
+    if [[ -n "\$found_file1" ]]; then
+        file_path=${file_path1}
+    elif [[ -n "\$found_file2" ]]; then
+        file_path=${file_path2}
+    elif [[ -n "\$found_file3" ]]; then
+        file_path=${file_path3}
+    else
+        echo "No matching PTUSE observations found for ${target_name} at ${utc_start}. Exiting."
+        exit 0
     fi
 
     # Extracting basename and filename without extension from ephemeris_file
@@ -336,18 +411,17 @@ process dspsr_fold_ephemeris_ptuse {
     
 
     # Loop through all files in the specified path
-    for input_file in \$(ls -v ${file_path}); do
+    for input_file in \$(ls -v \${file_path}); do
         # Extract the filename without its path and extension
         filename=\$(basename "\$input_file")
         filename_no_ext="\${filename%.*}"
 
-        
-        output_filename="\${filename_no_ext}.ar"
+        output_filename="\${filename_no_ext}_\${ephemeris_name_no_ext}"
 
         # Process each file with dspsr. If dspsr fails for a file, the script continues without exiting or throwing an error.
-        if dspsr -scloffs -A -E ${ephemeris_file} -L ${subint_length} -b ${bins} -t ${threads} -k ${telescope} -e ar -O "\${filename_no_ext}" "\$input_file" >& /dev/null; then
+        if dspsr -scloffs -A -E ${ephemeris_file} -L ${subint_length} -b ${bins} -t ${threads} -k ${telescope} -e ar -O "\${output_filename}" "\$input_file" >& /dev/null; then
             # If dspsr succeeds, add the path of the output .ar file to the list of successful files
-            successful_files+=("\$output_filename")
+            successful_files+=("\${output_filename}.ar")
         fi
 
         
@@ -356,9 +430,13 @@ process dspsr_fold_ephemeris_ptuse {
     # Check if there are any successful files to process further
     if [[ \${#successful_files[@]} -gt 0 ]]; then
         # Construct the psradd command using the array of successfully processed files
-        psradd -o "${target_name}_${utc_start}_\${ephemeris_name_no_ext}.ar" \${successful_files[@]}
+        psradd -o "${target_name}_${utc_start}_\${ephemeris_name_no_ext}.ar" *.ar
+
         for file in "\${successful_files[@]}"; do
-            rm -rfv "\$file"
+            #This check is for the edge case when user requested subint length is smaller than tobs, dspsr runs successfully, but no outputfile is created!
+            if [ -e "\$file" ]; then
+                rm -rfv "\$file"
+            fi
         done
     else
         echo "No files were successfully processed by dspsr."
